requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# LoRA
peft>=0.4.0
safetensors>=0.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Evaluation Metrics
pytorch-fid>=0.3.0
lpips>=0.1.4
torchmetrics>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Demo
gradio>=3.50.0

# Utils
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
PyYAML>=6.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Data
datasets>=2.14.0

